change Integer.BYTES to Float.BYTES in #readFloat

+++ src/main/java/lightletter/LetterContent.java
@@ -66,7 +66,7 @@ public class LetterContent extends LightLetter {
 	public float readFloat() {
 		checkBounds(Float.BYTES);
 		float f = PrimitiveReadUtil.readFloat(data, marker);
-		advance(Integer.BYTES);
+		advance(Float.BYTES);
 		return f;
 	}
